# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommunityResponse:
    properties:
      address:
        type: string
      branch_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.CreateCommunityRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      description:
        type: string
      name:
        type: string
      organization_id:
        type: integer
    required:
    - address
    - branch_id
    - name
    - organization_id
    type: object
  dto.CreateUnitRequest:
    properties:
      bathrooms:
        type: number
      branch_id:
        type: string
      description:
        type: string
      organization_id:
        type: integer
      rooms:
        type: integer
      square_feet:
        type: integer
      unit_number:
        type: string
    required:
    - bathrooms
    - branch_id
    - organization_id
    - rooms
    - square_feet
    - unit_number
    type: object
...
  dto.ListCommunitiesResponse:
    properties:
      communities:
        items:
          $ref: '#/definitions/dto.CommunityResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListUnitsResponse:
    properties:
      total:
        type: integer
      units:
        items:
          $ref: '#/definitions/dto.UnitResponse'
        type: array
    type: object
...
  dto.UnitResponse:
    properties:
      bathrooms:
        type: number
      branch_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      organization_id:
        type: integer
      rooms:
        type: integer
      square_feet:
        type: integer
      unit_number:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateCommunityRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      description:
        type: string
      name:
        type: string
      organization_id:
        type: integer
    type: object
  dto.UpdateUnitRequest:
    properties:
      bathrooms:
        type: number
      branch_id:
        type: string
      description:
        type: string
      organization_id:
        type: integer
      rooms:
        type: integer
      square_feet:
        type: integer
      unit_number:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a template API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Template API
  version: "1.0"
paths:
...
  /communities:
    get:
      description: Retrieve a list of communities with pagination and sorting support.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 'Sort fields (format: field1:asc,field2:desc) - Available fields:
          first_name, last_name, email, created_at'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of communities with pagination information
          schema:
            $ref: '#/definitions/dto.ListCommunitiesResponse'
        "400":
          description: Invalid input parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List communities
      tags:
      - community
    post:
      consumes:
      - application/json
      description: Create a new community with the provided details.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Community information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommunityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created community
          schema:
            $ref: '#/definitions/dto.CommunityResponse'
        "400":
          description: Invalid input, bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create community
      tags:
      - community
  /communities/{id}:
    delete:
      description: Delete a community by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Community not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete community
      tags:
      - community
    get:
      description: Retrieve a community's details using its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Community details
          schema:
            $ref: '#/definitions/dto.CommunityResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Community not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get community by ID
      tags:
      - community
    put:
      consumes:
      - application/json
      description: Update a community's details by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Community information to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommunityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated community details
          schema:
            $ref: '#/definitions/dto.CommunityResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Community not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update community
      tags:
      - community
  /communities/search:
    get:
      description: Search for a community by name or email with optional case sensitivity,
        pagination, and sorting.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Search query (e.g., community name )
        in: query
        name: search
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Case sensitive search (default: false)'
        in: query
        name: case_sensitive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of communities matching search criteria with pagination
            info
          schema:
            $ref: '#/definitions/dto.ListCommunitiesResponse'
        "400":
          description: Invalid input parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search community
      tags:
      - community
...
  /units:
    get:
      description: Retrieve a list of units with pagination and sorting support.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 'Sort fields (format: field1:asc,field2:desc) - Available fields:
          unit_number, name, created_at'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of units with pagination information
          schema:
            $ref: '#/definitions/dto.ListUnitsResponse'
        "400":
          description: Invalid input parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List units
      tags:
      - unit
    post:
      consumes:
      - application/json
      description: Create a new unit with the provided details.
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Unit information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created unit
          schema:
            $ref: '#/definitions/dto.UnitResponse'
        "400":
          description: Invalid input, bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create unit
      tags:
      - unit
  /units/{id}:
    delete:
      description: Delete a unit by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Unit not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete unit
      tags:
      - unit
    get:
      description: Retrieve a unit's details using its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unit details
          schema:
            $ref: '#/definitions/dto.UnitResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Unit not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get unit by ID
      tags:
      - unit
    put:
      consumes:
      - application/json
      description: Update a unit's details by its ID
      parameters:
      - description: Organization ID
        in: header
        name: Organization-ID
        required: true
        type: string
      - description: Branch ID
        in: header
        name: Branch-ID
        required: true
        type: string
      - description: Unit ID
        in: path
        name: id
        required: true
        type: string
      - description: Unit information to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Unit details
          schema:
            $ref: '#/definitions/dto.UnitResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Unit not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update unit
      tags:
      - unit
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"